+++ examples/regression_synthetic/main.go
@@ -17,19 +17,19 @@ func main() {
 	// 1. Generate synthetic data (8 features)
 	numSamples := 1000
 	dim := 8
-	inputs := make([][]float64, numSamples)
-	targets := make([][]float64, numSamples)
+	inputs := make([][]float32, numSamples)
+	targets := make([][]float32, numSamples)
 
 	// Simple linear combination with some non-linearity and noise
 	for i := 0; i < numSamples; i++ {
-		inputs[i] = make([]float64, dim)
-		sum := 0.0
+		inputs[i] = make([]float32, dim)
+		sum := float32(0.0)
 		for j := 0; j < dim; j++ {
-			inputs[i][j] = rand.Float64()*2 - 1
-			sum += inputs[i][j] * float64(j+1)
+			inputs[i][j] = rand.Float32()*2 - 1
+			sum += inputs[i][j] * float32(j+1)
 		}
 		// Target = sum + sum^2/10 + noise
-		targets[i] = []float64{sum + (sum*sum)/10.0 + rand.NormFloat64()*0.1}
+		targets[i] = []float32{sum + (sum*sum)/10.0 + float32(rand.NormFloat64()*0.1)}
 	}
 
 	// 2. Create architecture (8 -> 64 -> 32 -> 1)
@@ -47,12 +47,12 @@ func main() {
 	fmt.Printf("Training for %d epochs...\n", epochs)
 
 	for epoch := 1; epoch <= epochs; epoch++ {
-		totalLoss := 0.0
+		totalLoss := float32(0.0)
 		for i := range inputs {
 			totalLoss += network.Train(inputs[i], targets[i])
 		}
 		if epoch%20 == 0 {
-			fmt.Printf("Epoch %d - Avg Loss: %.6f\n", epoch, totalLoss/float64(numSamples))
+			fmt.Printf("Epoch %d - Avg Loss: %.6f\n", epoch, totalLoss/float32(numSamples))
 		}
 	}
