+++ examples/sequential_mnist/test_dense/test_dense_only.go
@@ -46,7 +46,7 @@ func downloadFile(url, dest string) error {
 	return err
 }
 
-func loadMNISTImages(filename string) ([][]float64, int, int, error) {
+func loadMNISTImages(filename string) ([][]float32, int, int, error) {
 	file, err := os.Open(filename)
 	if err != nil {
 		return nil, 0, 0, err
@@ -65,21 +65,21 @@ func loadMNISTImages(filename string) ([][]float64, int, int, error) {
 	if magic != 2051 {
 		return nil, 0, 0, fmt.Errorf("invalid magic number: %d", magic)
 	}
-	images := make([][]float64, numImages)
+	images := make([][]float32, numImages)
 	pixelCount := int(rows * cols)
 	for i := 0; i < int(numImages); i++ {
 		pixels := make([]uint8, pixelCount)
 		gz.Read(pixels)
-		floatPixels := make([]float64, pixelCount)
+		floatPixels := make([]float32, pixelCount)
 		for j := 0; j < pixelCount; j++ {
-			floatPixels[j] = float64(pixels[j]) / 255.0
+			floatPixels[j] = float32(pixels[j]) / 255.0
 		}
 		images[i] = floatPixels
 	}
 	return images, int(rows), int(cols), nil
 }
 
-func loadMNISTLabels(filename string) ([][]float64, error) {
+func loadMNISTLabels(filename string) ([][]float32, error) {
 	file, err := os.Open(filename)
 	if err != nil {
 		return nil, err
@@ -96,18 +96,18 @@ func loadMNISTLabels(filename string) ([][]float64, error) {
 	if magic != 2049 {
 		return nil, fmt.Errorf("invalid magic number: %d", magic)
 	}
-	labels := make([][]float64, numLabels)
+	labels := make([][]float32, numLabels)
 	for i := 0; i < int(numLabels); i++ {
 		var label uint8
 		binary.Read(gz, binary.BigEndian, &label)
-		oneHot := make([]float64, 10)
+		oneHot := make([]float32, 10)
 		oneHot[label] = 1.0
 		labels[i] = oneHot
 	}
 	return labels, nil
 }
 
-func argmax(v []float64) int {
+func argmax(v []float32) int {
 	maxIdx := 0
 	maxVal := v[0]
 	for i := 1; i < len(v); i++ {
@@ -173,6 +173,6 @@ func main() {
 			correct++
 		}
 	}
-	accuracy := float64(correct) / float64(len(xTest))
+	accuracy := float32(correct) / float32(len(xTest))
 	fmt.Printf("Test Accuracy: %.2f%%\n", accuracy*100)
 }
